GetStringSize returns fixed.Int26_6 now to avoid loss of precision.
The GetStringSize func used to return ints, which worked fine when there
was only one text chunk.  The cumulative loss in precision from
converting fixed.Int26_6 to ints, over multiple chunks, would clip a few
pixels on the right side.

The right-side clipping may still occur because we do lose some
precision in converting the fixed.Int26_6s to ints, but it should be rare.